Remove old search service
NHUB-572
