+++ newsroom/news_api/news/search_service.py
@@ -1,669 +1,123 @@
-import json
-import functools
-import re
-from dateutil import parser
-import pytz
-from datetime import datetime
+from typing import Any, cast
+from urllib.parse import urlencode
 
-from bson import ObjectId
-from werkzeug.datastructures import MultiDict
+from superdesk.flask import request as flask_request
+from superdesk.core.types.web import Request, Response
 
-from superdesk.core import get_app_config
-from superdesk import get_resource_service
-from superdesk.utc import utcnow, local_to_utc
-from superdesk.errors import SuperdeskApiError
-from content_api.errors import BadParameterValueError, UnexpectedParameterError
-
+from newsroom.types import SectionEnum
+from newsroom.types.wire import WireItem
+from newsroom.wire.service import WireItemService
 from newsroom.auth.utils import get_company_or_none_from_request
-from newsroom.wire.items import ItemsResource
-from newsroom.news_api.utils import (
-    post_api_audit,
-    remove_internal_renditions,
-    check_association_permission,
+from newsroom.products.utils import get_products_by_company_async
+from newsroom.search.types import NewshubSearchRequest, SearchFilterFunction
+from newsroom.search.base_service import BaseNewshubSearchService
+from newsroom.news_api.utils import check_association_permission, post_api_audit, remove_internal_renditions
+from newsroom.search.filters import apply_company_filter, apply_section_filter, apply_products_filter
+
+from .filters import (
+    apply_date_filter,
+    apply_filter_fields,
+    apply_projection,
+    apply_request_filter,
+    prefill_company,
+    prefill_products,
+    validate_page,
 )
-from newsroom.search.service import BaseSearchService, query_string
-from newsroom.products.products import get_products_by_company
+from .types import NewsApiSearchRequestArgs
 
 
-class NewsAPINewsService(BaseSearchService):
-    # set of parameters that the API will allow.
-    allowed_params = {
-        "start_date",  # ** alias=created_from
-        "end_date",  # ** alias=created_to
-        "include_fields",
-        "exclude_fields",
-        "page_size",  # ** alias=size
-        "page",  # ** alias=from
-        "version",
-        "where",  # ? Should we support this in original API?
-        "q",  # ** in original
-        "default_operator",  # ** in original
-        "filter",  # ** in original
-        "service",
-        "subject",
-        "genre",
-        "urgency",
-        "priority",
-        "type",
-        "item_source",
-        "sort",  # ** in original
-        "timezone",  # alias=timezone_offset?? Double check this one
-        "products",  # alias=product (this version support list of products ID, might also support that in original)
-    }
+default_search_filters = [
+    prefill_company,
+    prefill_products,
+    apply_section_filter,
+    apply_company_filter,
+    apply_products_filter,
+    apply_filter_fields,
+    apply_date_filter,
+    apply_request_filter,
+    apply_projection,
+    validate_page,
+]
 
-    default_sort = [{"versioncreated": "desc"}]
 
-    # set of fields that are allowed to be excluded in the exlude_fields parameter
-    allowed_exclude_fields = {
-        "version",
-        "versioncreated",
-        "firstcreated",
-        "headline",
-        "byline",
-        "slugline",
-    }
+class NewsApiSearchServiceAsync(BaseNewshubSearchService[NewsApiSearchRequestArgs, WireItem]):
+    search_args_class = NewsApiSearchRequestArgs
+    filters = cast(list[SearchFilterFunction], default_search_filters)
+    section = SectionEnum.NEWS_API
+    default_sort = [{"versioncreated", 1}]
+    default_page_size = 25
 
-    # set of fields that can be specified in the include_fields parameter
-    allowed_include_fields = {
-        "type",
-        "urgency",
-        "priority",
-        "language",
-        "description_html",
-        "located",
-        "keywords",
-        "source",
-        "subject",
-        "place",
-        "wordcount",
-        "charcount",
-        "body_html",
-        "readtime",
-        "profile",
-        "service",
-        "genre",
-        "associations",
-    }
+    def __init__(self):
+        self.service = WireItemService()
 
-    default_fields = {
-        "_id",
-        "uri",
-        "embargoed",
-        "pubstatus",
-        "ednote",
-        "signal",
-        "copyrightnotice",
-        "copyrightholder",
-        "versioncreated",
-        "evolvedfrom",
-        "original_id",
-    }
+    async def process_web_request(self, request):
+        resp = await super().process_web_request(request)
 
-    # set of fields that will be removed from all responses, we are not currently supporting associations and
-    # the products embedded in the items are the superdesk products
-    mandatory_exclude_fields = {"_current_version", "products"}
-
-    section = "news_api"
-    limit_days_setting = "news_api_time_limit_days"
-
-    def get(self, req, lookup):
-        resp = super().get(req, lookup)
-
-        orig_request_params = getattr(req, "args", MultiDict())
-
-        # Can't get the exclude projection to work do pop the exclude fields here
-        exclude_fields = (
-            self.mandatory_exclude_fields.union(set(orig_request_params.get("exclude_fields").split(",")))
-            if orig_request_params.get("exclude_fields")
-            else self.mandatory_exclude_fields
-        )
-
-        company = get_company_or_none_from_request(None)
-        # TODO-ASYNC: replace when NewsAPI is async
-        products = get_products_by_company(
-            company.to_dict(context={"use_objectid": True}) if company else None, product_type="news_api"
-        )
-
-        for doc in resp.docs:
-            for field in exclude_fields:
-                doc.pop(field, None)
-
-            if "associations" in orig_request_params.get("include_fields", ""):
-                if not check_association_permission(doc, products):
-                    doc.pop("associations", None)
-                else:
-                    remove_internal_renditions(doc)
+        await self.process_post_api_audit(resp)
+        await self.process_response_enhancements(request, resp)
 
         return resp
 
-    def prefill_search_query(self, search, req=None, lookup=None):
-        """Generate the search query instance
+    async def process_post_api_audit(self, response: Response):
+        # TODO-ASYNC: adjust once post_audit is async
+        post_api_audit(response.body)
 
-        :param newsroom.search.SearchQuery search: The search query instance
-        :param ParsedRequest req: The parsed in request instance from the endpoint
-        :param dict lookup: The parsed in lookup dictionary from the endpoint
-        """
+    async def process_response_enhancements(self, request: Request, response: Response):
+        search_req = self.get_search_request_instance(request)
+        self.build_hateoas(request, response, search_req)
 
-        self.prefill_search_args(search, req)
-        self.prefill_search_lookup(search, lookup)
-        self.prefill_search_page(search)
-        self.prefill_search_company(search)
-        self.prefill_search_section(search)
-        self.prefill_search_products(search)
-        self.prefill_search_items(search)
+        company = get_company_or_none_from_request(request)
+        assert company is not None
+        products = [
+            product.to_dict()
+            for product in await get_products_by_company_async(company, product_type=SectionEnum.NEWS_API)
+        ]
 
-    def apply_filters(self, search):
-        """Generate and apply the different search filters
+        for doc in response.body["_items"] or []:
+            self._enhance_internal_item_hateoas(doc)
 
-        :param newsroom.search.SearchQuery search: the search query instance
-        """
+            if "associations" in (search_req.args.include_fields or []):
+                self._check_associations(doc, products)
 
-        self.apply_section_filter(search)
-        self.apply_company_filter(search)
-        self.apply_products_filter(search)
-        self.apply_fields_filter(search)
-        self.apply_date_filter(search)
-        self.apply_request_filter(search)
-        self.apply_projections(search)
-
-        if len(search.query["bool"].get("should", [])):
-            search.query["bool"]["minimum_should_match"] = 1
-
-    def prefill_search_company(self, search):
-        company = get_company_or_none_from_request(None)
-        search.company = company.to_dict(context={"use_objectid": True}) if company else None
-
-    def prefill_search_products(self, search):
-        """Prefill the search products
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        if search.args.get("products"):
-            search.products = list(
-                get_resource_service("products").get(
-                    req=None,
-                    lookup={
-                        "is_enabled": True,
-                        "companies": ObjectId(search.company["_id"]),
-                        "_id": {"$in": [ObjectId(pid) for pid in search.args["products"].split(",")]},
-                    },
-                )
-            )
+    def _check_associations(self, doc: dict[str, Any], products: list[dict[str, Any]]):
+        if not check_association_permission(doc, products):
+            doc.pop("associations", None)
         else:
-            # TODO-ASYNC: replace when NewsAPI is async
-            search.products = get_products_by_company(search.company, product_type=search.section)
+            remove_internal_renditions(doc)
 
-    def prefill_search_args(self, search, req=None):
-        """Prefill the search request args
+    def build_hateoas(self, req: Request, resp: Response, search_req: NewshubSearchRequest[NewsApiSearchRequestArgs]):
+        base_url = req.path.strip("/")
+        query_params = search_req.args.to_dict()
 
-        :param newsroom.search.SearchQuery search: The search query instance
-        :param ParsedRequest req: The passed in request instance
-        """
+        resp.body.setdefault("_links", {})
+        resp.body["_links"]["parent"] = {"title": "Home", "href": "/"}
 
-        if req is None:
-            search.args = {}
-        elif getattr(req.args, "to_dict", None):
-            search.args = req.args.to_dict()
-        elif isinstance(req.args, dict):
-            search.args = req.args
-        else:
-            search.args = {}
+        # append page and page_size only if they were provided in original request
+        for arg in ["page", "page_size"]:
+            if arg in flask_request.view_args:
+                query_params.update({arg: query_params.get(arg)})
 
-        search.req = req
+        q_args = f"?{urlencode(query_params)}" if query_params else ""
+        resp.body["_links"]["self"] = (
+            {
+                "title": "News Search",
+                "href": f"{base_url}{q_args}",
+            },
+        )
 
-        # Validating the supplied arguments here, before we pre-fill other arguments
-        self.validate_include_exclude_fields(search)
-        self.validate_unknown_params(search, whitelist=self.allowed_params)
+        # add next page if there are more items
+        if (search_req.args.page * search_req.args.page_size) < resp.body["_meta"]["total"]:
+            query_params["page"] = search_req.args.page + 1
+            resp.body["_links"]["next"] = {"title": "next page", "href": f"{base_url}?{urlencode(query_params)}"}
 
-    def prefill_search_page(self, search):
-        """Prefill the search page parameters
+        # add prev page if needed
+        if search_req.args.page > 1:
+            query_params["page"] = search_req.args.page - 1
+            resp.body["_links"]["prev"] = {"title": "prev page", "href": f"{base_url}?{urlencode(query_params)}"}
 
-        :param newsroom.search.SearchQuery search: The search query instance
-        :param ParsedRequest req: The parsed in request instance from the endpoint
-        """
-
-        try:
-            search.args["page"] = int(search.args.get("page") or 1)
-        except ValueError:
-            raise BadParameterValueError("Page must be a number")
-
-        try:
-            search.args["page_size"] = int(search.args.get("page_size") or 25)
-        except ValueError:
-            raise BadParameterValueError("Page Size must be a number")
-
-        search.args["size"] = search.args["page_size"]
-        search.args["from"] = (search.args["page"] - 1) * search.args["page_size"]
-
-        super().prefill_search_page(search)
-
-        if isinstance(search.args["sort"], str):
-            if search.args["sort"] == "versioncreated:desc":
-                search.args["sort"] = [{"versioncreated": "desc"}]
-            elif search.args["sort"] == "versioncreated:asc":
-                search.args["sort"] = [{"versioncreated": "asc"}]
-            elif search.args["sort"] == "score":
-                search.args["sort"] = [{"_score": "desc"}]
-            else:
-                raise BadParameterValueError("Unknown sort option ({name})".format(name=search.args["sort"]))
-
-    def validate_request(self, search):
-        """Validate the request parameters
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        if not search.company:
-            raise SuperdeskApiError.forbiddenError()
-
-        self.validate_page(search)
-
-    def validate_include_exclude_fields(self, search):
-        """Validate the include/exclude fields param
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-        if search.args.get("include_fields") and not all(
-            p in [c for c in list(self.allowed_include_fields)] for p in search.args["include_fields"].split(",")
-        ):
-            raise UnexpectedParameterError("Include fields contains a non-allowed value")
-        elif search.args.get("exclude_fields") and not all(
-            p in [c for c in list(self.allowed_exclude_fields)] for p in search.args["exclude_fields"].split(",")
-        ):
-            raise UnexpectedParameterError("Exclude fields contains a non-allowed value")
-
-    def validate_page(self, search):
-        """Validate the page params
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-        query_max_page_size = get_app_config("QUERY_MAX_PAGE_SIZE")
-        if search.args["page"] < 1:
-            raise BadParameterValueError("Page number must be greater or equal to 1")
-        elif search.args["page_size"] > query_max_page_size:
-            raise BadParameterValueError(
-                "Requested maximum number of results exceeds {max}".format(max=query_max_page_size)
-            )
-        elif (search.args["page"] - 1) * search.args["page_size"] >= 1000:
-            # https://www.elastic.co/guide/en/elasticsearch/guide/current/pagination.html#pagination
-            raise BadParameterValueError("Page limit exceeded")
-
-    def validate_unknown_params(self, search, whitelist, allow_filtering=True):
-        """Check if the request contains only allowed parameters.
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        :param whitelist: iterable containing the names of allowed parameters.
-        :param bool allow_filtering: whether or not the filtering parameter is
-            allowed (True by default). Used for disallowing it when retrieving
-            a single object.
-
-        :raises UnexpectedParameterError:
-            * if the request contains a parameter that is not whitelisted
-            * if the request contains more than one value for any of the
-              parameters
-        """
-
-        arg_keys = search.args.keys()
-
-        if not allow_filtering:
-            err_msg = "Filtering{} is not supported when retrieving a " 'single object (the "{param}" parameter)'
-
-            if "start_date" in arg_keys:
-                desc = err_msg.format(" by date range", param="start_date")
-                raise UnexpectedParameterError(desc=desc)
-
-            if "end_date" in arg_keys:
-                desc = err_msg.format(" by date range", param="end_date")
-                raise UnexpectedParameterError(desc=desc)
-
-        for param_name in arg_keys:
-            if param_name not in whitelist:
-                raise UnexpectedParameterError(desc="Unexpected parameter ({})".format(param_name))
-
-            if len(search.req.args.getlist(param_name)) > 1:
-                desc = "Multiple values received for parameter ({})"
-                raise UnexpectedParameterError(desc=desc.format(param_name))
-
-    def apply_request_filter(self, search, highlights=True):
-        """Generate the filters from request args
-
-        :param newsroom.search.SearchQuery search:  The search query instance
-        """
-
-        if search.args.get("q"):
-            search.query["bool"]["filter"].append(
-                query_string(search.args["q"], search.args.get("default_operator") or "AND")
-            )
-
-    def apply_date_filter(self, search):
-        """Generate and apply date filters
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        start_date, end_date = self._get_date_range(search.args)
-        date_filter = self._create_date_range_filter(start_date, end_date)
-        if date_filter:
-            search.query["bool"]["filter"].append(date_filter)
-
-    def apply_fields_filter(self, search):
-        """Generate the field filters
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        # request argments and elasticsearch fields
-        argument_fields = {
-            "service": "service.code",
-            "subject": "subject.code",
-            "urgency": "urgency",
-            "priority": "priority",
-            "genre": "genre.code",
-            "item_source": "source",
+    def _enhance_internal_item_hateoas(self, item: dict[str, Any]):
+        item.setdefault("_links", {})
+        item["_links"]["self"] = {
+            "href": f"news/item/{item['_id']}",
+            "title": "News Item",
         }
-
-        filters = []
-
-        if search.args.get("filter"):
-            try:
-                filters = json.loads(search.args["filter"])
-            except Exception:
-                raise BadParameterValueError("Bad parameter value for Parameter (filter)")
-
-        for argument_name, field_name in argument_fields.items():
-            if argument_name not in search.args or search.args[argument_name] is None:
-                continue
-
-            filter_value = search.args[argument_name]
-
-            try:
-                filter_value = json.loads(filter_value)
-            except Exception:
-                pass
-
-            if not filter_value:
-                raise BadParameterValueError("Bad parameter value for Parameter ({})".format(argument_name))
-
-            if not isinstance(filter_value, list):
-                filter_value = [filter_value]
-
-            filters.append({"terms": {field_name: filter_value}})
-
-        if filters:
-            search.query["bool"]["filter"].extend(filters)
-
-    def apply_projections(self, search):
-        """Generate and apply projections
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        include_fields, exclude_fields = self._get_field_filter_params(search.args)
-        projection = self._create_field_filter(include_fields, exclude_fields)
-        search.projections = json.dumps(projection)
-
-    def gen_source_from_search(self, search):
-        """Generate the eve source object from the search query instance
-
-        :param newsroom.search.SearchQuery search: The search query instance
-        """
-
-        search.args["aggs"] = False
-        super().gen_source_from_search(search)
-
-    def get_internal_request(self, search):
-        """Creates an eve internal request object
-
-        :param newsroom.search.SearchQuery search: the search query instance
-        :return:
-        """
-
-        internal_req = super().get_internal_request(search)
-        del internal_req.projection
-
-        if search.args.get("df"):
-            internal_req["df"] = search.args["df"]
-
-        return internal_req
-
-    def _get_field_filter_params(self, request_params):
-        """Extract the list of content fields to keep in or remove from
-        the response.
-
-        The parameter names are `include_fields` and `exclude_fields`. Both are
-        simple comma-separated lists, for example::
-
-            exclude_fields=  field_1, field_2,field_3,, ,field_4,
-
-        NOTE: any redundant spaces, empty field values and duplicate values are
-        gracefully ignored and do not cause an error.
-
-        :param dict request_params: request parameter names and their
-            corresponding values
-
-        :return: a (include_fields, exclude_fields) tuple with each item being
-            either a `set` of field names (as strings) or None if the request
-            does not contain the corresponding parameter
-
-        :raises BadParameterValueError:
-            * if the request contains both parameters at the same time
-            * if any of the parameters contain an unknown field name (i.e. not
-                defined in the resource schema)
-            * if `exclude_params` parameter contains a field name that is
-                required to be present in the response according to the NINJS
-                standard
-        """
-        include_fields = request_params.get("include_fields")
-        exclude_fields = request_params.get("exclude_fields")
-
-        # parse field filter parameters...
-        strip_items = functools.partial(map, lambda s: s.strip())
-        remove_empty = functools.partial(filter, None)
-
-        if include_fields is not None:
-            include_fields = include_fields.split(",")
-            include_fields = set(remove_empty(strip_items(include_fields)))
-
-        if exclude_fields is not None:
-            exclude_fields = exclude_fields.split(",")
-            exclude_fields = set(remove_empty(strip_items(exclude_fields)))
-
-        # check for semantically incorrect field filter values...
-        if (include_fields is not None) and (exclude_fields is not None):
-            err_msg = "Cannot both include and exclude content fields " "at the same time."
-            raise UnexpectedParameterError(desc=err_msg)
-
-        if include_fields is not None:
-            err_msg = "Unknown content field to include ({})."
-            for field in include_fields:
-                if field not in ItemsResource.schema:
-                    raise BadParameterValueError(desc=err_msg.format(field))
-
-        if exclude_fields is not None:
-            if "uri" in exclude_fields:
-                err_msg = "Cannot exclude a content field required by the " "NINJS format (uri)."
-                raise BadParameterValueError(desc=err_msg)
-
-            err_msg = "Unknown content field to exclude ({})."
-            for field in exclude_fields:
-                if field not in ItemsResource.schema:
-                    raise BadParameterValueError(desc=err_msg.format(field))
-
-        if request_params.get("include_fields"):
-            include_fields = self.default_fields.union(include_fields)
-        else:
-            include_fields = self.default_fields.union(self.allowed_exclude_fields)
-
-        return include_fields, exclude_fields
-
-    def _create_field_filter(self, include_fields, exclude_fields):
-        """Create an `Eve` projection object that explicitly includes/excludes
-        particular content fields from results.
-
-        At least one of the parameters *must* be None. The created projection
-        uses either a whitlist or a blacklist approach (see below), it cannot
-        use both at the same time.
-
-        * If `include_fields` is not None, a blacklist approach is used. All
-            fields are _omittted_ from the result, except for those listed in
-            the `include_fields` set.
-        * If `exclude_fields` is not None, a whitelist approach is used. All
-            fields are _included_ in the result, except for those listed in the
-            `exclude_fields` set.
-        * If both parameters are None, no field filtering should be applied
-        and an empty dictionary is returned.
-
-        NOTE: fields required by the NINJS standard are _always_ included in
-        the result, regardless of the field filtering parameters.
-
-        :param include_fields: fields to explicitly include in result
-        :type include_fields: set of strings or None
-        :param exclude_fields: fields to explicitly exclude from result
-        :type exclude_fields: set of strings or None
-
-        :return: `Eve` projection filter (as a dictionary)
-        """
-        projection = {}
-
-        if include_fields is not None:
-            for field in include_fields:
-                projection[field] = 1
-        #        elif exclude_fields is not None:
-        #            for field in exclude_fields:
-        #                projection[field] = 0
-
-        return projection
-
-    @staticmethod
-    def _parse_iso_date(date_str, timezone=None):
-        """Create a date object from the given string in ISO 8601 format.
-
-        :param date_str:
-        :type date_str: str or None
-
-        :return: resulting date object or None if None is given
-        :rtype: datetime.date
-
-        :raises ValueError: if `date_str` is not in the ISO 8601 date format
-        """
-        if date_str is None:
-            return None
-        else:
-            dt = parser.parse(date_str)
-            if dt.tzinfo is None:
-                if timezone:
-                    if timezone not in pytz.all_timezones:
-                        raise BadParameterValueError("Bad parameter value for Parameter (timezone)")
-                    dt = local_to_utc(timezone, dt)
-                else:
-                    dt = pytz.timezone("UTC").localize(dt)
-            return dt
-
-    def _get_date_range(self, request_params):
-        """Extract the start and end date limits from request parameters.
-
-        If start and/or end date parameter is not present, a default value is
-        returned for the missing parameter(s).
-
-        :param dict request_params: request parameter names and their
-            corresponding values
-
-        :return: a (start_date, end_date) tuple with both values being
-            instances of Python's datetime.date
-
-        :raises BadParameterValueError:
-            * if any of the dates is not in the ISO 8601 format
-            * if any of the dates is set in the future
-            * if the start date is bigger than the end date
-        """
-
-        # regex that should match likely relative elastic searcg date math
-        regex = r"now([-+][0-9]*([YMwdHhms]*)$|/d$)"
-
-        # check date limits' format...
-        err_msg = "{} parameter must be a valid ISO 8601 date (YYYY-MM-DD) " "with optional the time part"
-
-        relative_start = False
-        relative_end = False
-        try:
-            # check for a relative date
-            if re.match(regex, request_params.get("start_date", "")):
-                start_date = request_params.get("start_date")
-                relative_start = True
-            else:
-                start_date = self._parse_iso_date(request_params.get("start_date"), request_params.get("timezone"))
-        except BadParameterValueError:
-            raise
-        except ValueError:
-            raise BadParameterValueError(desc=err_msg.format("start_date")) from None
-
-        try:
-            if request_params.get("end_date"):
-                if re.match(regex, request_params.get("end_date", "")):
-                    end_date = request_params.get("end_date")
-                    relative_end = True
-                else:
-                    end_date = self._parse_iso_date(request_params.get("end_date"), request_params.get("timezone"))
-            else:
-                end_date = None
-        except ValueError:
-            raise BadParameterValueError(desc=err_msg.format("end_date")) from None
-
-        # disallow dates in the future...
-        err_msg = "{} date ({}) must not be set in the future " "(current server date (UTC): {})"
-        today = utcnow()
-
-        if (start_date is not None) and not relative_start and (start_date > today):
-            raise BadParameterValueError(desc=err_msg.format("Start", start_date.isoformat(), today.isoformat()))
-
-        if (end_date is not None) and not relative_end and (end_date > today):
-            raise BadParameterValueError(desc=err_msg.format("End", end_date.isoformat(), today.isoformat()))
-
-        # make sure that the date range limits make sense...
-        if (
-            (not relative_start or not relative_end)
-            and (start_date is not None)
-            and (end_date is not None)
-            and (start_date > end_date)
-        ):
-            # NOTE: we allow start_date == end_date (for specific date queries)
-            raise BadParameterValueError(desc="Start date must not be greater than end date")
-
-        return start_date, end_date
-
-    def _create_date_range_filter(self, start_date, end_date):
-        """Create a MongoDB date range query filter from the given dates.
-
-        If both the start date and the end date are None, an empty filter is
-        returned. The filtering is performed on the `versioncreated` field in
-        database.
-
-        :param start_date: the minimum version creation date (inclusive)
-        :type start_date: datetime.date or None
-        :param end_date: the maximum version creation date (inclusive)
-        :type end_date: datetime.date or None
-
-        :return: MongoDB date range filter (as a dictionary)
-        """
-        if (start_date is None) and (end_date is None):
-            return {}  # nothing to set for the date range filter
-
-        if end_date is None:
-            end_date = "now"
-
-        return {
-            "range": {
-                "versioncreated": {
-                    "gte": start_date if isinstance(start_date, str) else self._format_date(start_date),
-                    "lte": end_date if isinstance(end_date, str) else self._format_date(end_date),
-                }
-            }
-        }
-
-    @staticmethod
-    def _format_date(date):
-        return datetime.strftime(date, get_app_config("ELASTIC_DATETIME_FORMAT"))
-
-    def on_fetched(self, doc):
-        post_api_audit(doc)
